Resolve "[Server] Clients movement in local network"

Closes #63

See merge request arzarv/iti0200-2020-project!39

+++ core/src/com/multimage/network/packets/FirstPacket.java
@@ -5,7 +5,7 @@ import com.multimage.MultiMage;
 public class FirstPacket {
 
     public int id = -1;
-    public int x = (int) (200 / MultiMage.PPM);
-    public int y = (int) (50 / MultiMage.PPM);
+    public float x = (200 / MultiMage.PPM);
+    public float y = (50 / MultiMage.PPM);
 
 }
