+++ core/src/com/multimage/network/packets/Position.java
@@ -3,7 +3,7 @@ package com.multimage.network.packets;
 public class Position {
 
     public int playerID;
-    public int posX;
-    public int posY;
+    public float posX;
+    public float posY;
 
 }
